refactor: Board 컴포넌트와 Game 컴포넌트를 분리
